fix potential double free
double free could occur if patch allocated the dst_data, then an
error occured later on, causing the jump to fail, which frees the dst_data.
however because dst_data is a **, then the user will have a non-null pointer
so they may assume it should be freed.

patch will always free and NULL the dst_data (if allocated) on error.
only on success will dst_data remain allocated.